move test class and class to the good package and remove duplicate test class
