+++ app/src/main/java/io/github/polysmee/profile/MainUserInfoDataStore.java
@@ -1,4 +1,4 @@
-package io.github.polysmee.settings;
+package io.github.polysmee.profile;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
